// src/Components/Statisticks/types.ts
export type TTabsStatistics =
  | "daily"
  | "conversions"
  | "affiliates"
  | "advertisers"
  | "offers"
  | "sales-managers"
  | "affiliate-managers"
  | "referrals"
  | "countries"
  | "cities"
  | "os"
  | "goals"
  | "devices"
  | "mobile-isp"
  | "connection-type"
  | "landing"
  | "affiliate-postbacks"
  | "server-postbacks"
  | "retention-rate"
  | "kpi"
  | "categories"
  | "user-agents"
  | "custom"
  | "comparison-report";